sLLM/requirements.txt
ollama==0.1.2
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
kafka-python==2.0.2
elasticsearch==8.10.0
numpy==1.26.1
pandas==2.1.2
scikit-learn==1.5.0
transformers==4.34.1
torch==2.2.0
python-dotenv==1.0.0
schedule==1.2.0